integrate RuleSetConfig in CalciteConfig
